move the component to a better directory since its not a pure CDS component
